+++ src/main/java/servlets/eventPoints/EventPointsListServlet.java
@@ -1,6 +1,5 @@
-package servlets;
+package servlets.eventPoints;
 
-import models.Festival;
 import services.EventPoinServiceImpl;
 
 import javax.servlet.RequestDispatcher;
@@ -10,27 +9,16 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.List;
 
-@WebServlet("/eventpoints")
+@WebServlet("/eventpoints/list")
 public class EventPointsListServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 
         List eventPoints = EventPoinServiceImpl.getInstance().getAllList();
         request.setAttribute("EventPointsList", eventPoints);
-        RequestDispatcher dispatcher = request.getRequestDispatcher("/eventPointsList.jsp");
+        RequestDispatcher dispatcher = request.getRequestDispatcher("/eventPoints/eventPointsList.jsp");
         dispatcher.forward(request, response);
-    }
-
-    @Override
-    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        String idString =request.getParameter("deleteId");
-        System.out.println(idString);
-        long id = Long.parseLong(idString);
-        EventPoinServiceImpl.getInstance().remove(id);
-
-        response.sendRedirect("/eventpoints");
     }
 }
